+++ src/components/Block/View.jsx
@@ -30,6 +30,9 @@ const View = (props) => {
   const [confirmFacebook, setConfirmFacebook] = useState(
     !!Number(cookies.confirm_facebook),
   );
+  const [confirmGoogle, setConfirmGoogle] = useState(
+    !!Number(cookies.confirm_google),
+  );
 
   const expiryDate = new Date();
   expiryDate.setMonth(expiryDate.getMonth() + 1);
@@ -59,6 +62,12 @@ const View = (props) => {
       setCookie('confirm_youtube', 1, options);
     } else {
       removeCookie('confirm_youtube', options);
+    }
+
+    if (confirmGoogle) {
+      setCookie('confirm_google', 1, options);
+    } else {
+      removeCookie('confirm_google', options);
     }
 
     setCookie('confirm_cookies', 1, options);
@@ -103,6 +112,16 @@ const View = (props) => {
               label="Facebook"
               onChange={() => setConfirmFacebook(!confirmFacebook)}
               checked={confirmFacebook}
+            />
+          </Form.Field>
+        )}
+        {includes(modules, 'google') && (
+          <Form.Field>
+            <Checkbox
+              toggle
+              label="Google"
+              onChange={() => setConfirmGoogle(!confirmGoogle)}
+              checked={confirmGoogle}
             />
           </Form.Field>
         )}